Get tasks from DB and take them into account after parse

+++ app/src/main/java/net/planner/planetapp/database/TaskDB.kt
@@ -42,6 +42,6 @@ class TaskDB(
     val maxDivisionsNumber: Int = 10,
 
     @SerializedName("durationInMinutes")
-    val durationInMinutes: Long = 60,
+    val durationInMinutes: Int = 60
 
 )
